test: change testfile layout, use one parser for Reforged and Netease
